# crates/secure_smoke_service/openapi.yaml
openapi: "3.1.0"
info:
  title: SunLit Security Smoke-Test Service
  description: >
    A purpose-built axum microservice with 54 routes, each exercising a
    specific security control against a known attack class. Includes 15
    mobile security routes covering MASVS controls. Designed for automated
    DAST scanning with OWASP ZAP.
  version: "0.1.0"
  license:
    name: MIT
servers:
  - url: http://127.0.0.1:3001
    description: Local development server
paths:
...
  /smoke/mobile/cert-pin:
    post:
      summary: Certificate pin verification (MASVS-NETWORK-1)
      operationId: smokeMobileCertPin
      tags: [mobile-network]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CertPinRequest"
      responses:
        "200":
          description: Pin valid
        "422":
          description: Pin mismatch or invalid hash
  /smoke/mobile/cleartext:
    post:
      summary: Cleartext traffic detection (MASVS-NETWORK-2)
      operationId: smokeMobileCleartext
      tags: [mobile-network]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CleartextRequest"
      responses:
        "200":
          description: URL is secure
        "422":
          description: Cleartext traffic blocked
  /smoke/mobile/storage-policy:
    post:
      summary: Mobile storage policy enforcement (MASVS-STORAGE-1)
      operationId: smokeMobileStoragePolicy
      tags: [mobile-storage]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StoragePolicyRequest"
      responses:
        "200":
          description: Storage policy compliant
        "422":
          description: Storage policy violated
  /smoke/mobile/sensitive-buffer:
    post:
      summary: Sensitive buffer handling (MASVS-STORAGE-1)
      operationId: smokeMobileSensitiveBuffer
      tags: [mobile-storage]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SensitiveBufferRequest"
      responses:
        "200":
          description: Buffer handled securely
  /smoke/mobile/biometric:
    post:
      summary: Biometric result validation (MASVS-AUTH-2)
      operationId: smokeMobileBiometric
      tags: [mobile-auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BiometricRequest"
      responses:
        "200":
          description: Biometric accepted
        "422":
          description: Biometric rejected
  /smoke/mobile/step-up:
    post:
      summary: Step-up authentication enforcement (MASVS-AUTH-3)
      operationId: smokeMobileStepUp
      tags: [mobile-auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StepUpRequest"
      responses:
        "200":
          description: Step-up not required
        "422":
          description: Step-up required
  /smoke/mobile/deep-link:
    post:
      summary: Deep link URL validation (MASVS-PLATFORM-1)
      operationId: smokeMobileDeepLink
      tags: [mobile-platform]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeepLinkRequest"
      responses:
        "200":
          description: Deep link valid
        "422":
          description: Deep link rejected
  /smoke/mobile/webview-url:
    post:
      summary: WebView URL safety check (MASVS-PLATFORM-1)
      operationId: smokeMobileWebviewUrl
      tags: [mobile-platform]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WebViewUrlRequest"
      responses:
        "200":
          description: WebView URL safe
        "422":
          description: WebView URL blocked
  /smoke/mobile/clipboard:
    post:
      summary: Clipboard security policy check (MASVS-PLATFORM-2)
      operationId: smokeMobileClipboard
      tags: [mobile-platform]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClipboardRequest"
      responses:
        "200":
          description: Clipboard policy evaluated
  /smoke/mobile/root-detect:
    post:
      summary: Environment detection signal processing (MASVS-RESILIENCE-1)
      operationId: smokeMobileRootDetect
      tags: [mobile-resilience]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RootDetectRequest"
      responses:
        "200":
          description: RASP decision returned
  /smoke/mobile/app-integrity:
    post:
      summary: App integrity verification (MASVS-RESILIENCE-2)
      operationId: smokeMobileAppIntegrity
      tags: [mobile-resilience]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppIntegrityRequest"
      responses:
        "200":
          description: Integrity valid
        "422":
          description: Integrity check failed
  /smoke/mobile/pii-classify:
    post:
      summary: PII classification (MASVS-PRIVACY-1)
      operationId: smokeMobilePiiClassify
      tags: [mobile-privacy]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PiiClassifyRequest"
      responses:
        "200":
          description: PII classification result
  /smoke/mobile/pseudonymize:
    post:
      summary: Data pseudonymization (MASVS-PRIVACY-2)
      operationId: smokeMobilePseudonymize
      tags: [mobile-privacy]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PseudonymizeRequest"
      responses:
        "200":
          description: Pseudonymized value returned
  /smoke/mobile/consent:
    post:
      summary: Consent policy enforcement (MASVS-PRIVACY-3)
      operationId: smokeMobileConsent
      tags: [mobile-privacy]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConsentRequest"
      responses:
        "200":
          description: Consent allowed
        "422":
          description: Consent denied or not collected
...